Move WindowManager to its own file
